Move value error and error codes out of values mod
`values/mod.rs` file is too lengthy.
